ids-multipart: refactor package structure for multipart client
