+++ data-protocols/ids/ids-api-multipart-dispatcher-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/client/message/MultipartResponse.java
@@ -1,4 +1,4 @@
-package org.eclipse.dataspaceconnector.ids.api.multipart.client;
+package org.eclipse.dataspaceconnector.ids.api.multipart.client.message;
 
 import de.fraunhofer.iais.eis.Message;
